requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
blis==0.7.11
brotlicffi @ file:///C:/miniconda3/conda-bld/brotlicffi_1762950484957/work
cachetools==6.2.2
catalogue==2.0.10
certifi @ file:///C:/miniconda3/conda-bld/certifi_1763135349059/work/certifi
cffi @ file:///C:/miniconda3/conda-bld/cffi_1761832792265/work
charset-normalizer @ file:///C:/miniconda3/conda-bld/charset-normalizer_1761745005832/work
click==8.3.1
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.3
confection==0.1.5
cymem==2.0.13
dataclasses-json==0.6.7
datasets==4.4.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
executing==2.2.1
faiss==1.8.0
fastjsonschema==2.21.2
filelock @ file:///C:/miniconda3/conda-bld/filelock_1762419922635/work
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna @ file:///C:/miniconda3/conda-bld/idna_1761912044050/work
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
langchain==1.0.8
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.7
langchain-openai==1.0.3
langchain-text-splitters==1.0.0
langcodes==3.5.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.44
language_data==1.3.0
lark==1.3.1
llama_cpp_python==0.3.16
marisa-trie==1.3.1
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mistune==3.1.4
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.15
mypy_extensions==1.1.0
narwhals==2.12.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
neo4j==5.21.0
nest-asyncio==1.6.0
networkx==3.2
notebook==7.5.0
notebook_shim==0.2.4
numpy @ file:///D:/bld/numpy_1707225528937/work/dist/numpy-1.26.4-cp311-cp311-win_amd64.whl#sha256=537f0573fefb1e93b3633e8a79102e6b6d2911e09e0a90008993efef4242448e
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.5
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
platformdirs==4.5.0
preshed==3.0.12
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pure_eval==0.2.3
pyarrow==21.0.0
pycparser @ file:///C:/miniconda3/conda-bld/pycparser_1757496060165/work
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pysbd==0.3.4
PySocks @ file:///C:/miniconda3/conda-bld/pysocks_1761753028158/work
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
pywinpty==3.0.2
PyYAML @ file:///C:/miniconda3/conda-bld/pyyaml_1763465512847/work
pyzmq==27.1.0
ragas==0.1.4
referencing==0.37.0
regex==2025.11.3
requests @ file:///C:/miniconda3/conda-bld/requests_1762359611410/work
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.29.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
Send2Trash==1.8.3
sentence-transformers==2.7.0
six==1.17.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.2
stack-data==0.6.3
streamlit==1.51.0
sympy==1.13.1
tenacity==9.1.2
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.19.1
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.0
typer==0.9.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.14
weasel==0.3.4
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
win_inet_pton @ file:///C:/miniconda3/conda-bld/win_inet_pton_1761746278197/work
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
